LayoutTests/imported/w3c/web-platform-tests/background-fetch/match.https.window-expected.txt


FAIL Matching to a single request should work assert_equals: expected "Background Fetch" but got ""
PASS Matching to a non-existing request should work
FAIL Matching multiple times on the same request works as expected. assert_equals: expected "Background Fetch" but got ""
FAIL Access to active fetches is supported. assert_equals: expected "Background Fetch" but got ""
FAIL Match with query options. assert_equals: expected 2 but got 3
